Moved executable code from SparseMatrix.hpp to .cpp
Description:
To facilitate not recompiling the code in the header every time
its included, a lot of the executable code was moved to a source
file.

Other source files and CMake targets were updated accordingly.

Also moved the "using namespace std;" out of global scope in
SparseMatrix.hpp. This should have been a seperate commit (oops!).