Modifying configure to only install pre-commit hook if needed
Description:
Previously, the configure steps to install pre-commit hook were
run every time the project was configured. This may have caused
an issue with the python cache when a subsequent commit was attempted.

In any case, the pre-commit hook is only installed during configure
if it has not already been installed in the repo.